Allow images on borders
Our documentation standards require our Confluence images to have
borders around them. There is not currently a way to do this, but adding
it to the processer is not complex.

This adds support for a `border=true` attribute to the image macro which
emits an `ac:border="true` attribute on the resulting `ac:image` tag.