Allow exteding the classpath of the docker image
Allow providing new extension (read custom macros) in the
/opt/extensions folder inside the container. This allows extensions to
be compiled and then mounted into the publisher.

This also changes the 'asciidoc-confluence-publisher-docker' jar from being a
runnable jar, to regular jar. This change was made so there's only one
place that has to be updated (in the docker modeule) if the main class
name ever changes.

Closes: #460